Refatorando tela de relatórios e também ajustando eixo Y para alinhar com a altura do gráfico
