+++ app/src/main/java/com/recuperavc/data/db/AppRoomDatabase.kt
@@ -50,7 +50,7 @@ abstract class AppRoomDatabase : RoomDatabase() {
                     AppRoomDatabase::class.java,
                     "app.db"
                 )
-                    //.fallbackToDestructiveMigration() // <- só ativar isso aqui para ele limpar sozinho tudo do db se mudar o schema
+                    .fallbackToDestructiveMigration() // <- só ativar isso aqui para ele limpar sozinho tudo do db se mudar o schema
                     .build()
                     .also { INSTANCE = it }
             }
