+++ app/src/main/java/com/recuperavc/ui/main/ReportTypes.kt
@@ -1,0 +1,6 @@
+package com.recuperavc.ui.main
+
+enum class ChartType { WPM, WER }
+enum class CoherenceChartType { TIME, ERRORS }
+enum class ReportTab { Audio, Coherence, Motion }
+
