+++ app/src/main/java/com/recuperavc/ui/main/components/SegmentedTabs.kt
@@ -1,0 +1,103 @@
+package com.recuperavc.ui.main.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Gesture
+import androidx.compose.material.icons.filled.Mic
+import androidx.compose.material.icons.filled.TrendingUp
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.recuperavc.ui.main.ReportTab
+import com.recuperavc.ui.theme.GreenDark
+import com.recuperavc.ui.theme.GreenLight
+
+@Composable
+fun SegmentedTabs(tab: ReportTab, onTab: (ReportTab) -> Unit) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(20.dp))
+            .background(GreenLight.copy(alpha = 0.2f))
+            .padding(6.dp),
+        horizontalArrangement = Arrangement.spacedBy(6.dp)
+    ) {
+        SegButton(
+            modifier = Modifier.weight(1f),
+            selected = tab == ReportTab.Audio,
+            icon = Icons.Default.Mic,
+            text = "Voz",
+            onClick = { onTab(ReportTab.Audio) }
+        )
+        SegButton(
+            modifier = Modifier.weight(1f),
+            selected = tab == ReportTab.Coherence,
+            icon = Icons.Default.TrendingUp,
+            text = "Raciocínio",
+            onClick = { onTab(ReportTab.Coherence) }
+        )
+        SegButton(
+            modifier = Modifier.weight(1f),
+            selected = tab == ReportTab.Motion,
+            icon = Icons.Default.Gesture,
+            text = "Coordenação",
+            onClick = { onTab(ReportTab.Motion) }
+        )
+    }
+}
+
+@Composable
+private fun SegButton(
+    modifier: Modifier = Modifier,
+    selected: Boolean,
+    icon: ImageVector,
+    text: String,
+    onClick: () -> Unit
+) {
+    val bg = if (selected) GreenDark else Color.White
+    val fg = if (selected) Color.White else GreenDark
+    val elevation = if (selected) 4.dp else 0.dp
+
+    Card(
+        modifier = modifier,
+        shape = RoundedCornerShape(16.dp),
+        colors = CardDefaults.cardColors(containerColor = bg),
+        elevation = CardDefaults.cardElevation(defaultElevation = elevation)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .clickable { onClick() }
+                .padding(vertical = 14.dp, horizontal = 8.dp),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Icon(
+                icon,
+                contentDescription = null,
+                tint = fg,
+                modifier = Modifier.size(26.dp)
+            )
+            Spacer(Modifier.height(6.dp))
+            Text(
+                text,
+                color = fg,
+                fontWeight = FontWeight.Bold,
+                fontSize = 13.sp,
+                textAlign = TextAlign.Center
+            )
+        }
+    }
+}
+
