+++ app/src/main/java/com/recuperavc/ui/main/filters/DateFilterCard.kt
@@ -1,0 +1,160 @@
+package com.recuperavc.ui.main.filters
+
+import android.app.DatePickerDialog
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.CalendarToday
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.recuperavc.ui.theme.GreenDark
+import java.time.Instant
+import java.time.LocalDateTime
+import java.time.ZoneId
+import java.time.format.DateTimeFormatter
+import java.util.Calendar
+
+@Composable
+fun DateFilterCard(
+    startDate: Instant,
+    endDate: Instant,
+    isManuallySet: Boolean,
+    onStartDateChange: (Instant) -> Unit,
+    onEndDateChange: (Instant) -> Unit,
+    onTapSound: () -> Unit
+) {
+    val context = LocalContext.current
+
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        shape = RoundedCornerShape(16.dp),
+        colors = CardDefaults.cardColors(containerColor = Color.White),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
+    ) {
+        Column(modifier = Modifier.fillMaxWidth().padding(16.dp)) {
+            Text(
+                text = "Filtro de Data",
+                fontWeight = FontWeight.Bold,
+                fontSize = 18.sp,
+                color = Color.Black
+            )
+            Spacer(Modifier.height(8.dp))
+            if (!isManuallySet) {
+                Text(
+                    text = "Mostrando dados dos últimos 7 dias",
+                    fontSize = 14.sp,
+                    color = Color.Black.copy(alpha = 0.7f)
+                )
+                Spacer(Modifier.height(12.dp))
+            } else {
+                Spacer(Modifier.height(4.dp))
+            }
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.spacedBy(12.dp)
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+                    Text("Início", fontSize = 14.sp, fontWeight = FontWeight.SemiBold, color = Color.Black)
+                    Spacer(Modifier.height(4.dp))
+                    Row(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .clip(RoundedCornerShape(8.dp))
+                            .background(Color(0xFFF5F5F5))
+                            .clickable {
+                                onTapSound()
+                                val cal = Calendar.getInstance()
+                                val localDateTime = LocalDateTime.ofInstant(startDate, ZoneId.systemDefault())
+                                cal.set(localDateTime.year, localDateTime.monthValue - 1, localDateTime.dayOfMonth)
+
+                                DatePickerDialog(
+                                    context,
+                                    { _, year, month, day ->
+                                        val selected = LocalDateTime.of(year, month + 1, day, 0, 0)
+                                        onStartDateChange(selected.atZone(ZoneId.systemDefault()).toInstant())
+                                    },
+                                    cal.get(Calendar.YEAR),
+                                    cal.get(Calendar.MONTH),
+                                    cal.get(Calendar.DAY_OF_MONTH)
+                                ).show()
+                            }
+                            .padding(12.dp),
+                        verticalAlignment = Alignment.CenterVertically,
+                        horizontalArrangement = Arrangement.SpaceBetween
+                    ) {
+                        Text(
+                            text = DateTimeFormatter.ofPattern("dd/MM/yyyy").format(
+                                LocalDateTime.ofInstant(startDate, ZoneId.systemDefault())
+                            ),
+                            fontSize = 16.sp,
+                            fontWeight = FontWeight.Bold,
+                            color = GreenDark
+                        )
+                        Icon(
+                            Icons.Default.CalendarToday,
+                            contentDescription = null,
+                            tint = GreenDark,
+                            modifier = Modifier.size(20.dp)
+                        )
+                    }
+                }
+                Column(modifier = Modifier.weight(1f)) {
+                    Text("Fim", fontSize = 14.sp, fontWeight = FontWeight.SemiBold, color = Color.Black)
+                    Spacer(Modifier.height(4.dp))
+                    Row(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .clip(RoundedCornerShape(8.dp))
+                            .background(Color(0xFFF5F5F5))
+                            .clickable {
+                                onTapSound()
+                                val cal = Calendar.getInstance()
+                                val localDateTime = LocalDateTime.ofInstant(endDate, ZoneId.systemDefault())
+                                cal.set(localDateTime.year, localDateTime.monthValue - 1, localDateTime.dayOfMonth)
+
+                                DatePickerDialog(
+                                    context,
+                                    { _, year, month, day ->
+                                        val selected = LocalDateTime.of(year, month + 1, day, 23, 59, 59)
+                                        onEndDateChange(selected.atZone(ZoneId.systemDefault()).toInstant())
+                                    },
+                                    cal.get(Calendar.YEAR),
+                                    cal.get(Calendar.MONTH),
+                                    cal.get(Calendar.DAY_OF_MONTH)
+                                ).show()
+                            }
+                            .padding(12.dp),
+                        verticalAlignment = Alignment.CenterVertically,
+                        horizontalArrangement = Arrangement.SpaceBetween
+                    ) {
+                        Text(
+                            text = DateTimeFormatter.ofPattern("dd/MM/yyyy").format(
+                                LocalDateTime.ofInstant(endDate, ZoneId.systemDefault())
+                            ),
+                            fontSize = 16.sp,
+                            fontWeight = FontWeight.Bold,
+                            color = GreenDark
+                        )
+                        Icon(
+                            Icons.Default.CalendarToday,
+                            contentDescription = null,
+                            tint = GreenDark,
+                            modifier = Modifier.size(20.dp)
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
+
